requirements.txt
Django~=5.2.2
pillow
django-cleanup
django-allauth~=65.9.0
django-htmx
channels~=4.2.2
asgiref~=3.8.1
shortuuid~=1.0.13
dj-database-url
django-environ